issue 4 done

+++ app/src/main/java/org/jik/notification_proto/adapter/KeywordAdapter.kt
@@ -1,6 +1,8 @@
 package org.jik.notification_proto.adapter
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -36,26 +38,34 @@ class KeywordAdapter(var keywords: List<KeywordEntity>,var onDeleteListener: OnD
 
         holder.keywordname.text = keyword.keyword
         holder.keyworddelete.setOnClickListener {
+            // 다이얼로그로 키워드 삭제를 할 지 물어본다.
+            val builder = AlertDialog.Builder(holder.itemView.context)
+            builder.setTitle("키워드를 삭제하시겠습니까?")
+                    .setPositiveButton("확인", DialogInterface.OnClickListener { dialogInterface, i ->
+                        // 키워드 삭제 내용을 서버로 전달
+                        // val token = this.getContext().getSharedPreferences("token", Context.MODE_PRIVATE)?.getString("token","default value")
+                        var deletedata = DeleteModel(token, keyword = keyword.keyword)
+                        Log.d("deletedata", deletedata.toString())
+                        APIS.create().delete_users(deletedata).enqueue(object : Callback<String> {
+                            override fun onResponse(call: Call<String>, response: Response<String>) {
+                                Log.d("log", response.toString())
+                                Log.d("log", response.body().toString())
+                                // 서버응답이 200 일 때(네트워크가 연결되었을 때, 서버가 켜져있을 때)
+                                onDeleteListener.onDeleteListener(keyword)
+                            }
 
+                            override fun onFailure(call: Call<String>, t: Throwable) {
+                                Log.d("log", t.printStackTrace().toString())
+                                Log.d("log", "fail")
+                                Toast.makeText(holder.itemView.context,"네트워크 연결을 확인 해주세요! ", Toast.LENGTH_SHORT).show()
+                            }
+                        })
+                    })
+                    .setNegativeButton("취소", DialogInterface.OnClickListener { dialogInterface, i ->
+                        Log.d("키워드 삭제 취소","키워드 삭제를 취소하였습니다.")
+                    })
+            builder.show()
 
-            // 키워드 삭제 내용을 서버로 전달
-            // val token = this.getContext().getSharedPreferences("token", Context.MODE_PRIVATE)?.getString("token","default value")
-            var deletedata = DeleteModel(token, keyword = keyword.keyword)
-            Log.d("deletedata", deletedata.toString())
-            APIS.create().delete_users(deletedata).enqueue(object : Callback<String> {
-                override fun onResponse(call: Call<String>, response: Response<String>) {
-                    Log.d("log", response.toString())
-                    Log.d("log", response.body().toString())
-                    // 서버응답이 200 일 때(네트워크가 연결되었을 때, 서버가 켜져있을 때)
-                    onDeleteListener.onDeleteListener(keyword)
-                }
-
-                override fun onFailure(call: Call<String>, t: Throwable) {
-                    Log.d("log", t.printStackTrace().toString())
-                    Log.d("log", "fail")
-                    Toast.makeText(holder.itemView.context,"네트워크 연결을 확인 해주세요! ", Toast.LENGTH_SHORT).show()
-                }
-            })
         }
     }
